fix: move RunAllTests into non _test.go
Allows other packages to run RunAllTests in their unit tests
