+++ tests/check/check.go
@@ -7,6 +7,7 @@ import (
 	v1parser "github.com/craigpastro/openfga-dsl-parser"
 	parser "github.com/craigpastro/openfga-dsl-parser/v2"
 	"github.com/openfga/openfga/assets"
+	"github.com/openfga/openfga/pkg/tuple"
 	"github.com/openfga/openfga/pkg/typesystem"
 	"github.com/openfga/openfga/tests"
 	"github.com/stretchr/testify/require"
@@ -38,20 +39,79 @@ type assertion struct {
 	Trace            string
 }
 
-type CheckTestClientInterface interface {
+// ClientInterface defines client interface for running check tests
+type ClientInterface interface {
 	tests.TestClientBootstrapper
 	Check(ctx context.Context, in *openfgapb.CheckRequest, opts ...grpc.CallOption) (*openfgapb.CheckResponse, error)
 }
 
-func runSchema1_1CheckTests(t *testing.T, client CheckTestClientInterface) {
+// RunAllTests will run all check tests
+func RunAllTests(t *testing.T, client ClientInterface) {
+	t.Run("RunAllTests", func(t *testing.T) {
+		t.Run("Check", func(t *testing.T) {
+			t.Parallel()
+			testCheck(t, client)
+		})
+		t.Run("BadAuthModelID", func(t *testing.T) {
+			t.Parallel()
+			testBadAuthModelID(t, client)
+		})
+	})
+}
+
+func testCheck(t *testing.T, client ClientInterface) {
+	t.Run("Schema1_1", func(t *testing.T) {
+		t.Parallel()
+		runSchema1_1CheckTests(t, client)
+	})
+	t.Run("Schema1_0", func(t *testing.T) {
+		t.Parallel()
+		runSchema1_0CheckTests(t, client)
+	})
+}
+
+func testBadAuthModelID(t *testing.T, client ClientInterface) {
+
+	ctx := context.Background()
+	resp, err := client.CreateStore(ctx, &openfgapb.CreateStoreRequest{Name: "bad auth id"})
+	require.NoError(t, err)
+
+	storeID := resp.GetId()
+	model := `
+	type user
+
+	type doc
+	  relations
+	    define viewer: [user] as self
+	    define can_view as viewer
+	`
+	_, err = client.WriteAuthorizationModel(ctx, &openfgapb.WriteAuthorizationModelRequest{
+		StoreId:         storeID,
+		SchemaVersion:   typesystem.SchemaVersion1_1,
+		TypeDefinitions: parser.MustParse(model),
+	})
+	require.NoError(t, err)
+	_, err = client.Check(ctx, &openfgapb.CheckRequest{
+		StoreId:              storeID,
+		TupleKey:             tuple.NewTupleKey("doc:x", "viewer", "user:y"),
+		AuthorizationModelId: "01GS89AJC3R3PFQ9BNY5ZF6Q97",
+	})
+
+	require.Error(t, err)
+	e, ok := status.FromError(err)
+	require.True(t, ok)
+	require.Equal(t, int(openfgapb.ErrorCode_authorization_model_not_found), int(e.Code()))
+}
+
+func runSchema1_1CheckTests(t *testing.T, client ClientInterface) {
 	runTests(t, typesystem.SchemaVersion1_1, client)
 }
 
-func runSchema1_0CheckTests(t *testing.T, client CheckTestClientInterface) {
+func runSchema1_0CheckTests(t *testing.T, client ClientInterface) {
 	runTests(t, typesystem.SchemaVersion1_0, client)
 }
 
-func runTests(t *testing.T, schemaVersion string, client CheckTestClientInterface) {
+func runTests(t *testing.T, schemaVersion string, client ClientInterface) {
 	var b []byte
 	var err error
 	if schemaVersion == typesystem.SchemaVersion1_1 {
