+++ tests/writemodel/writemodel_test.go
@@ -28,12 +28,3 @@ func TestWriteAuthorizationModel(t *testing.T) {
 
 	RunAllTests(t, pb.NewOpenFGAServiceClient(conn))
 }
-
-func RunAllTests(t *testing.T, client ClientInterface) {
-	t.Run("RunAllTests", func(t *testing.T) {
-		t.Run("WriteTest", func(t *testing.T) {
-			t.Parallel()
-			runTests(t, client)
-		})
-	})
-}
